requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --no-emit-index-url --output-file=requirements.txt requirements.in
#
attrs==20.3.0
    # via
    #   -r requirements.in
    #   cattrs
    #   pytest
cattrs==1.3.0
    # via -r requirements.in
click==7.1.2
    # via flask
flask==1.1.2
    # via -r requirements.in
iniconfig==1.1.1
    # via pytest
itsdangerous==1.1.0
    # via flask
jinja2==2.11.3
    # via flask
markupsafe==1.1.1
    # via jinja2
packaging==20.9
    # via pytest
pluggy==0.13.1
    # via pytest
py==1.10.0
    # via pytest
pyparsing==2.4.7
    # via packaging
pytest==6.2.2
    # via -r requirements.in
toml==0.10.2
    # via pytest
werkzeug==1.0.1
    # via flask
